install.rs: Fix not compatible with some editors
Some editors such as `vscode` are exec by a bash script.

Child.wait() received a exit code from bash, but vscode is still running.

And the TempDir(`config.fetch.make_view()`) is delete when it goes out of scope.

Then, editor cannot found TempDir.

So, ask "Proceed with installation?" in it's scope.