Add an option to show upstream URLs
Added an option so that, when using looking for packages to install, they show their upstream URL.

Whenever I'm looking for a package, I like to see if they're updated regularly. So I go to the source to see when the last update was.

This option takes way the burden of doing `-Si` for every package. And since it's optional, if people don't want this, they can keep it turned off.