Force to recompute the signature
When rebuilding a package with a local repo, the signature is not
overwritten with a new one. This means that the new pkg is added to the
repo but the old signature is kept, which is invalid with the new build.

This change forces to overwrite the old signature with the new
recomputed one.